requirements/requirements.txt
pandas
scikit-learn
tensorflow
pyyaml
contractions
bs4
nltk
uvicorn
fastapi
transformers